+++ qcircuits/density_operator.py
@@ -14,6 +14,7 @@ aliased at the top-level module, so that one can call
 import numpy as np
 
 from qcircuits.operators import OperatorBase
+from itertools import chain
 
 
 class DensityOperator(OperatorBase):
@@ -42,6 +43,24 @@ class DensityOperator(OperatorBase):
         s += ' Tensor:\n'
         s += super().__str__()
         return s
+
+    @property
+    def probabilities(self):
+        """
+        Get the probability of observing each computational basis
+        vector upon making a measurement.
+
+        Returns
+        -------
+        numpy float64 multidimensional array
+            The probability associated with each computational basis
+            vector.
+        """
+
+        probs = np.real(np.diag(self.to_matrix())).reshape([2] * (self.rank // 2))
+        assert abs(np.sum(probs) - 1.0) < 1e-4, ('State probabilities'
+                                                 'do not sum to 1.')
+        return probs
 
     @staticmethod
     def from_ensemble(states, ps=None):
@@ -105,20 +124,80 @@ class DensityOperator(OperatorBase):
 
         return result.transpose(permutation)
 
-    def _measurement_probabilites(self, qubit_indices):
-        # Put non-measured qubits up front
+    def _reduced_tensor(self, retain_indices):
+        # Put traced out qubits up front
+        traced_indices = list(
+            set(range(self.rank // 2)) - set(retain_indices)
+        )
+        t = self._permuted_tensor(traced_indices + retain_indices)
+        d1 = len(traced_indices)
+        d2 = len(retain_indices)
+        D1 = 2**d1
+        idx = list(chain(range(0, 2 * d1, 2), range(1, 2 * d1, 2), range(2 * d1, self.rank)))
+        t = np.transpose(t, idx).reshape([D1] * 2 + [2] * 2 * d2)
+        return t[range(D1), range(D1), ...].sum(axis=0)
+
+    def _measurement_probabilities(self, qubit_indices):
         unmeasured_indices = list(
             set(range(self.rank // 2)) - set(qubit_indices)
         )
-        t = self._permuted_tensor(unmeasured_indices + qubit_indices)
 
-        # Trace out unmeasured bits
-        # TODO do this in a single operation
-        for _ in range(len(unmeasured_indices)):
-            t = np.sum(t[[0, 1], [0, 1], ...], axis=0)
-
+        t = self._reduced_tensor(qubit_indices)
         ps = np.real(np.diag(OperatorBase._tensor_to_matrix(t)))
         return ps, unmeasured_indices
+
+    def purify(self):
+        """
+        If this is a density operator for a :math:`d` qubit system, produce
+        a pure state for a :math:`2d` qubit system with the same measurement
+        probabilities for the first :math:`d` qubits, and whose reduced density
+        operator for the first :math:`d` qubits is equal to the original
+        density operator.
+
+        Returns
+        -------
+        State
+            A `2d` qubit pure state.
+        """
+
+        from qcircuits.state import State
+
+        eigvals, eigvecs = np.linalg.eig(self.to_matrix())
+        pure_column = (np.sqrt(np.expand_dims(eigvals, 0)) * eigvecs).flatten()
+        return State.from_column_vector(pure_column)
+
+    def reduced_density_operator(self, qubit_indices):
+        """
+        Compute the reduced density operator of the given qubits by tracing out
+        the qubits not given.
+
+        Parameters
+        ----------
+        qubit_indices : iterable
+            Indices indicating the qubit(s) that
+            we compute the reduced density operator of.
+
+        Returns
+        -------
+        DensityOperator
+            The reduced density operator of the sub-system.
+        """
+
+        if isinstance(qubit_indices, int):
+            qubit_indices = [qubit_indices]
+        qubit_indices = list(qubit_indices)
+
+        if qubit_indices == []:
+            raise ValueError('Must retain at least one qubit.')
+
+        if min(qubit_indices) < 0 or max(qubit_indices) >= self.rank // 2:
+            raise ValueError('Trying to measure qubit index i not 0<=i<d, '
+                             'where d is the rank of the state vector.')
+
+        if len(qubit_indices) != len(set(qubit_indices)):
+            raise ValueError('Qubit indices list contains repeated elements.')
+
+        return DensityOperator(self._reduced_tensor(qubit_indices))
 
     def measure(self, qubit_indices=None, remove=False):
         """
@@ -175,7 +254,7 @@ class DensityOperator(OperatorBase):
             raise ValueError('Qubit indices list contains repeated elements.')
 
         # Get measurement outcome probabilities
-        ps, unmeasured_indices = self._measurement_probabilites(qubit_indices)
+        ps, unmeasured_indices = self._measurement_probabilities(qubit_indices)
 
         num_outcomes = 2**len(qubit_indices)
         outcome = np.random.choice(num_outcomes, p=ps)